Stomped the memory corruption bug
It was my fault. I am an idiot, I have tried to set a cell outside an array,
generating a memory corruption error at runtime the first time I called
a new on the memory. The game tried to allocate the memory for a new sprite
and I suppose they store the memory status in a table, while the system
maintains information on where the application has written, so when C++
tries to allocate memory for a new class, it uses its own memory table
to know WHERE to allocate the memory.

In this case the new requires a memory location which already had informations
in it that the application did not know, but the operating system does.
As soon as this happens the OS will throw an error and signals the program to
stop.

Case solved!