SORRY PEOPLE! ANOTHER OVERHAUL!
Ok, so here is the deal: Cocos2d-x works using something they call
"autorelease" objects. I think they use it to manage the memory at runtime in
order to clean the memory once they do not need the object anymore, a
sort of Garbage Collector. In order to use this I have taken inspiration
from their implementation of the "Node::create" function and used it to
generate Grids and Taogtrominos.

I have understood slightly better the use of references and private
constructors. Now on with the game, dammit! XD